fix: move signals to separate module
