+++ apps/client/src/App.tsx
@@ -3,7 +3,7 @@ import { QueryClientProvider } from "@tanstack/solid-query";
 import { ErrorBoundary, Suspense } from "solid-js";
 import GamePage from "./pages/game/GamePage";
 import Layout from "./ui/Layout";
-import { authClient, queryClient } from "./api";
+import { authClient, queryClient } from "./state/api";
 import { ErrorFallback } from "./ui/ErrorFallback";
 import { Loading } from "./ui/Loading";
