+++ apps/client/src/game/AreaScene.tsx
@@ -4,7 +4,8 @@ import { useContext, createEffect, createMemo, Index } from "solid-js";
 import { EngineContext, Pixi } from "@mp/pixi/solid";
 import { createQuery } from "@tanstack/solid-query";
 import { loadTiledMapSpritesheets } from "@mp/tiled-renderer";
-import { api, myCharacter } from "../api";
+import { api } from "../state/api";
+import { myCharacter } from "../state/signals";
 import { CharacterActor } from "./CharacterActor";
 import { TileHighlight } from "./TileHighlight";
 import { getTilePosition } from "./getTilePosition";
